requirements.txt
pinecone-client>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=1.5.0
tqdm>=4.65.0
